Allow set env-vars per process-type
Setting environment vars prefixed with the process type + "__" (double
underscore chars) will be setted as a env-var to that specific process.

I.E:

- the app has 3 process: `beat`, `web` and `worker`
- we have as env-vars:
  - `WEB__MAGIC_NUMBER=2`
  - `WORKER__MAGIC_NUMBER=3`

Both env-vars will be read by the "deploy-util" script.

When it runs on the "web" unit, it will create the env-var
`MAGIC_NUMBER` with calue `2`.

When it runs on the "worker" unit, it will create the env-var
`MAGIC_NUMBER` with calue `3`.

When it runs on the "beat" unit, no env-var is set.

If there is a `MAGIC_NUMBER` is previous set on the app, it will be
overwriten by the more specific one.

This will only set env-vars to the process under the application
process. It means that accessing it usign `app-shell` will not process
it. If you want to load it, you could run:

```
./deploy-util.sh run_tsuru_app bash
```

(You could also replace `bash` to whatever you want to).

To test the process and/or see what is been setting, it is available
(this also does not change the current env, just shows the vars):

```
./deploy-util.sh set_env_vars_for_process
```

There is no automatic-tests on it, so there is some good test-cases:

- Values with white-spaces:
```
$ BEAT__COMPLEX__VAR__NAME='1 2 3 A' ./deploy-util.sh set_env_vars_for_process
Setting env-vars to beat process: COMPLEX__VAR__NAME='1 2 3 A'
```